third_party/blink/web_tests/accessibility/name-calc-aria-owns.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div class="container">
    <input id="input1" aria-labelledby="list1">
    <ul id="list1" aria-owns="list1_item3">
        <li>A
        <li>B
    </ul>
    <div role="listitem" id="list1_item3">C</div>
</div>

<script>
test(function(t) {
    console.log('aaaaaaaaaaaaa');
    var axInput1 = accessibilityController.accessibleElementById("input1");
    console.log('aaaaaaaaaaaaa 2');

    assert_equals(axInput1.name, "A B C");
}, "Aria-owns is considered in a name computation");
</script>
